ai-hugo-frontend/requirements.txt
streamlit==1.25.0
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
aiohttp==3.8.5
